Split up tests.
